legal-daily-brief/requirements.txt
# Legal Daily Brief — Dependencies
# pip install -r requirements.txt

msal==1.31.0               # Microsoft authentication
httpx==0.27.2              # HTTP client for Graph API calls
python-dotenv==1.0.1       # Load .env file
weasyprint==62.3           # HTML → PDF generation
                           # Mac: run  brew install pango  first
jinja2==3.1.4              # HTML templating
python-dateutil==2.9.0     # Date parsing from calendar events
loguru==0.7.2              # Clean logging
keyring==25.4.1            # Secure token storage — macOS Keychain on Mac,
                           # Windows Credential Manager on Windows
